Split processor detection out of path computation
