+++ subprojects/scala/src/test/groovy/org/gradle/api/tasks/scala/ScalaCompileTest.groovy
@@ -19,7 +19,7 @@ import org.gradle.api.InvalidUserDataException
 import org.gradle.api.file.FileCollection
 import org.gradle.api.internal.ConventionTask
 import org.gradle.api.internal.file.FileTreeInternal
-import org.gradle.api.internal.tasks.compile.AnnotationProcessorDetector
+import org.gradle.api.internal.tasks.compile.AnnotationProcessorPathFactory
 import org.gradle.api.internal.tasks.scala.ScalaJavaJointCompileSpec
 import org.gradle.api.tasks.TaskExecutionException
 import org.gradle.api.tasks.compile.AbstractCompile
@@ -34,7 +34,7 @@ class ScalaCompileTest extends AbstractCompileTest {
     private scalaCompiler = Mock(Compiler)
     private scalaClasspath = Mock(FileTreeInternal)
     private processorClasspath = Mock(FileCollection)
-    private processorDetector = Mock(AnnotationProcessorDetector)
+    private processorDetector = Mock(AnnotationProcessorPathFactory)
 
     @Override
     AbstractCompile getCompile() {
